Move Celery jobs into a subpackage.
Part of the fix for ZEN-32984.
