+++ Products/Zuul/facades/devicefacade.py
@@ -27,8 +27,7 @@ from Products.Zuul.facades import TreeFacade
 from Products.Zuul.catalog.component_catalog import get_component_field_spec, pad_numeric_values_for_indexing
 from Products.Zuul.catalog.interfaces import IModelCatalogTool
 from Products.Zuul.interfaces import IDeviceFacade, IInfo, ITemplateNode, IMetricServiceGraphDefinition
-from Products.Jobber.facade import FacadeMethodJob
-from Products.Jobber.jobs import SubprocessJob
+from Products.Jobber.jobs import FacadeMethodJob, SubprocessJob
 from Products.Zuul.tree import SearchResults
 from Products.DataCollector.Plugins import CoreImporter, PackImporter, loadPlugins
 from Products.ZenModel.DeviceOrganizer import DeviceOrganizer
